Changed config to allow self-hosted instances
No more .netrc file. Now the configuration file is under
'~/.config/habitica/auth.cfg', and contains the connection url with the
login and the password.

To avoid potential confusion, the old 'cache config file' is now named a
'cache file', and is under '~/.config/habitica/cache.cfg'. A section
'Quest' is used to keep the quest-related cached data.